games/tango_solver/requirements.txt
opencv-python
numpy
pillow
matplotlib
